requirements.txt
google-cloud-discoveryengine>=0.13.0
google-cloud-storage>=2.16.0
google-cloud-service-usage>=1.10.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
click>=8.1.7
rich>=13.7.0
